change data type short to int

+++ ToolsApi/Data/Models/ConnectionEvent.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public byte[] TransferId { get; set; } = null!;
         public ulong Offset { get; set; }
-        public ushort Objective { get; set; }
+        public uint Objective { get; set; }
         public ushort MimeType { get; set; }
     }
 }
